+++ native/ax/MpegDeMux-3/DeMuxFilter.h
@@ -299,7 +299,7 @@ private:
 
 	//for debug 
 	FILE* fd;
-	unsigned long offset;
+	size_t offset;
 	void open_dump();
 	void close_dump();
 	void rewind_dump( );
